rename: JwtToken 클래스명을 JwtReponse로 수정에 따른 변화
