+++ src/main/java/com/windstorm/management/implement/member/MemberLoginManager.java
@@ -1,7 +1,5 @@
 package com.windstorm.management.implement.member;
 
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
@@ -9,8 +7,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.windstorm.management.controller.auth.request.Login;
 import com.windstorm.management.domain.member.Member;
-import com.windstorm.management.security.jwt.JwtToken;
-import com.windstorm.management.security.jwt.JwtTokenProvider;
+import com.windstorm.management.implement.auth.AuthenticationGenerator;
+import com.windstorm.management.security.jwt.JwtProvider;
+import com.windstorm.management.security.jwt.JwtResponse;
 
 import lombok.RequiredArgsConstructor;
 
@@ -19,23 +18,20 @@ import lombok.RequiredArgsConstructor;
 public class MemberLoginManager {
 
 	private final MemberReader memberReader;
-	private final AuthenticationManager authenticationManager;
-	private final JwtTokenProvider jwtTokenProvider;
+	private final JwtProvider jwtProvider;
 	private final PasswordEncoder passwordEncoder;
+	private final AuthenticationGenerator authenticationGenerator;
 
 	@Transactional
-	public JwtToken login(Login request) {
+	public JwtResponse login(Login request) {
 		Member member = memberReader.read(request.uniqueMemberId());
 
 		if (!passwordEncoder.matches(request.password(), member.getPassword())) {
 			throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
 		}
 
-		UsernamePasswordAuthenticationToken authenticationToken =
-			new UsernamePasswordAuthenticationToken(request.uniqueMemberId(), request.password());
+		Authentication authentication = authenticationGenerator.authenticate(request);
 
-		Authentication authentication = authenticationManager.authenticate(authenticationToken);
-
-		return jwtTokenProvider.generateToken(authentication);
+		return jwtProvider.generateToken(authentication);
 	}
 }
