+++ src/main/java/com/windstorm/management/security/jwt/JwtAuthenticationFilter.java
@@ -16,16 +16,16 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class JwtAuthenticationFilter extends GenericFilterBean {
-	private final JwtTokenProvider jwtTokenProvider;
+	private final JwtProvider jwtProvider;
 
 	@Override
 	public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws
 		IOException, ServletException {
 		String token = resolveToken((HttpServletRequest)request);
 
-		if (token != null && jwtTokenProvider.validateToken(token)) {
+		if (token != null && jwtProvider.validateToken(token)) {
 			// 토큰이 유효할 경우 토큰에서 Authentication 객체를 가지고 와서 SecurityContext에 저장
-			Authentication authentication = jwtTokenProvider.getAuthentication(token);
+			Authentication authentication = jwtProvider.getAuthentication(token);
 			SecurityContextHolder.getContext().setAuthentication(authentication);
 		}
