+++ src/main/java/com/windstorm/management/security/jwt/JwtResponse.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 @Data
 @AllArgsConstructor
-public class JwtToken {
+public class JwtResponse {
 	private String grantType;
 	private String accessToken;
 	private String refreshToken;
